feat(geometry): move 25 source files to flui-geometry + flui-types bridge (U6+U7)
Per D-block PR-C-2 plan U6+U7 (squashed — U6 alone leaves flui-types broken
until U7 bridge lands; combined commit preserves green workspace per atomic-
commit-per-unit discipline).

U6 (move source):
- git mv crates/flui-types/src/geometry/{25 files} → crates/flui-geometry/src/
- Rename mod.rs → lib.rs (becomes flui-geometry crate root)
- rmdir crates/flui-types/src/geometry/
- Mechanical sed-rewrite crate::geometry:: → crate:: in 11 files
- Hand-fix grouped import in text_path.rs (crate::geometry::traits → crate::traits;
  sed pattern missed it because of {grouped} syntax)
- Add 'simd' feature to flui-geometry Cargo.toml (matrix4 references
  feature = "simd" for SSE/NEON gating; auto-detects target arch)
- Adopt geometry-specific lint allows (cast_precision_loss, many_single_char_names,
  similar_names, float_cmp, wildcard_imports, etc.) mirroring flui-types' pre-
  split Cargo.toml [lints.clippy] section — math/geometry code legitimately
  uses these patterns

U7 (flui-types bridge):
- crates/flui-types/Cargo.toml: add flui-geometry path dep
- crates/flui-types/src/lib.rs: replace 'pub mod geometry;' with
  'pub use flui_geometry as geometry;' — alias keeps flui_types::geometry::*
  consumers compiling unchanged. Direct flui_geometry::* paths also valid.

Top-level re-exports (flui_types::{EdgeInsets, Edges, Matrix4, Offset, Pixels,
Point, RRect, Rect, Size}) continue working via the bridge — no consumer
migration required in this commit. Downstream callers may migrate to
flui_geometry::* directly over time.

Per memo D9 file fences: PR-C-2 touchpoints crates/flui-geometry/src/* (new),
crates/flui-geometry/Cargo.toml (extended with simd feature + lint allows),
crates/flui-types/{Cargo.toml,src/lib.rs}.

Build gate: cargo build --workspace + cargo clippy --workspace --all-targets
-- -D warnings + bash scripts/port-check.sh -v all exit 0; cargo test on
flui-geometry/flui-types/flui-foundation/flui-view/flui-app all pass. (One
pre-existing flaky singleton-state test in flui-app test_semantics_enabled
flapped once mid-run; passes deterministically on isolated execution and on
re-run. Unrelated to geometry split — singleton RenderingFlutterBinding state
leakage between sibling tests.)
