+++ crates/flui-geometry/Cargo.toml
@@ -10,8 +10,42 @@ repository.workspace = true
 keywords = ["geometry", "graphics", "ui", "flui", "math"]
 categories = ["graphics", "mathematics"]
 
-[lints]
-workspace = true
+# Geometry-specific lint relaxations — mirrors crates/flui-types/Cargo.toml
+# pre-split. Math/geometry crates legitimately have many single-char names,
+# tolerated truncation casts, similar names (x/y/z), float comparisons (with
+# epsilon checks done at higher levels), and wildcard imports in preludes.
+[lints.rust]
+unsafe_code = "warn"
+rust_2018_idioms = "warn"
+
+[lints.clippy]
+# Inherit workspace-like base
+all = { level = "warn", priority = -1 }
+pedantic = { level = "warn", priority = -1 }
+# Workspace-level allows
+cast_precision_loss = "allow"
+module_name_repetitions = "allow"
+too_many_lines = "allow"
+missing_errors_doc = "allow"
+missing_panics_doc = "allow"
+must_use_candidate = "allow"
+# Math/geometry crate specifics
+many_single_char_names = "allow"
+cast_possible_truncation = "allow"
+cast_sign_loss = "allow"
+cast_lossless = "allow"
+similar_names = "allow"
+# Doc lints — will address incrementally
+doc_markdown = "allow"
+# Geometry code often passes small Copy types by ref for API consistency
+needless_pass_by_value = "allow"
+trivially_copy_pass_by_ref = "allow"
+# Builder-pattern methods returning Self are pervasive
+return_self_not_must_use = "allow"
+# Float comparisons are intentional in geometry/physics (tolerance-checked elsewhere)
+float_cmp = "allow"
+# Wildcard imports used in preludes
+wildcard_imports = "allow"
 
 [dependencies]
 # Numeric traits for generic math (Float, Num, etc.)
@@ -44,8 +78,12 @@ glam = ["dep:glam"]
 # mint interoperability
 mint = ["dep:mint"]
 
+# SIMD acceleration (auto-detects SSE/AVX on x86_64 or NEON on aarch64).
+# Gates SSE/NEON code paths in matrix4 multiplication; falls back to scalar.
+simd = []
+
 # Enable all features
-full = ["serde", "glam", "mint"]
+full = ["serde", "glam", "mint", "simd"]
 
 [package.metadata.docs.rs]
 all-features = true
