+++ crates/flui-geometry/src/corner.rs
