+++ crates/flui-geometry/src/edges.rs
