+++ crates/flui-geometry/src/error.rs
