+++ crates/flui-geometry/src/traits.rs
