+++ crates/flui-geometry/src/transform.rs
@@ -303,7 +303,7 @@ impl Transform {
     /// let t = Transform::rotate_radians(Radians::from_degrees(45.0));
     /// ```
     #[inline]
-    pub fn rotate_radians(angle: crate::geometry::Radians) -> Self {
+    pub fn rotate_radians(angle: crate::Radians) -> Self {
         Self::Rotate { angle: angle.0 }
     }
 
@@ -356,7 +356,7 @@ impl Transform {
     /// ```
     #[inline]
     pub fn rotate_around_radians(
-        angle: crate::geometry::Radians,
+        angle: crate::Radians,
         pivot_x: f32,
         pivot_y: f32,
     ) -> Self {
@@ -725,7 +725,7 @@ impl From<&Transform> for Matrix4 {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::geometry::units::px;
+    use crate::units::px;
 
     #[test]
     fn test_identity() {
